Made it compile without warning and run in 32-bits mode again, w.Python 2.4 and 2.5. Remains to test in 64 bit again...
git-svn-id: https://svn.code.sf.net/p/guppy-pe/code/trunk/guppy@7 a2a17790-8729-0410-921b-93eeb8b3d4a2

+++ src/heapy/horizon.c
@@ -52,7 +52,7 @@ horizon_remove(NyHorizonObject *v)
     }
     *p = v->next;
     if (!rm.horizons && rm.types) {
-	long i = 0;
+	Py_ssize_t i = 0;
 	PyObject *pk, *pv;
 	while (PyDict_Next(rm.types, &i, &pk, &pv)) {
 	    ((PyTypeObject *)pk)->tp_dealloc = (destructor) PyInt_AsLong(pv);
